% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F4.R
\name{BlockTrMean}
\alias{BlockTrMean}
\title{Mean of transposed Data block by block}
\usage{
BlockTrMean(
  Data = data(),
  lengthBlock = integer(),
  cls = c(),
  rs = c(),
  time_col = NULL,
  Mean = character()
)
}
\arguments{
\item{Data}{Data frame that will be imported.}

\item{lengthBlock}{Nomber of rows that will be transposed for any row.}

\item{cls}{Columns of the  data frame that we want to transpose.}

\item{rs}{The rows of the first block. Based on this rows the function will select identically the rows of every other block.}

\item{time_col}{The 'time parameter' is optional. We can ignore it if our data has no time values.}

\item{Mean}{Character that can take the values Row, Col or Block. Based on the option
that we'll choose and the function will return the mean of the Row, Col of each block or the general mean of all the block.}
}
\value{
tibble
}
\description{
Gives the rows mean or the columns mean of the blocks transposed data. It also gives the total mean of
each transposed block. give the mean of the column or of the lines of each block or of the
hole block transposed.

Gives the rows mean or the columns mean of the blocks transposed data. It also gives the total mean of
each transposed block.
}
