% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F5.R
\name{BlockTrSD}
\alias{BlockTrSD}
\title{SD of transposed Data block by block}
\usage{
BlockTrSD(
  Data = data(),
  lengthBlock = integer(),
  cls = c(),
  rs = c(),
  time_col = NULL,
  SD = character()
)
}
\arguments{
\item{Data}{Data frame that will be imported.}

\item{lengthBlock}{Nomber of rows that will be transposed for any row.}

\item{cls}{Columns of the  data frame that we want to transpose.}

\item{rs}{The rows of the first block. Based on this rows the function will select identically the rows of every other block.}

\item{time_col}{The 'time parameter' is optional. We can ignore it if our data has no time values.}

\item{SD}{Character that can take the values Row, Col or Block. Based on the option.
that we'll choose and the function will return the mean of the Row, Col of each block or the general mean of all the block.}
}
\value{
tibble
}
\description{
Takes any data fame and transpose the blocks predeterminated in the function, than it give
the standard deviation of the column or the standard deviation of the the rows of each block.
It also gives the total standard deviation of each transposed block. ' the hole block transposed.
}
