% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2.R
\name{InvDtBbyB}
\alias{InvDtBbyB}
\title{Inversing data frame block by block}
\usage{
InvDtBbyB(Data = data(), lengthBlock = integer(), cls = c(), rs = c())
}
\arguments{
\item{Data}{Data frame that will be imported}

\item{lengthBlock}{The length of the blocks that will be transposed.}

\item{cls}{The columns that we want to include in the transposition.}

\item{rs}{The rows of the first block. Based on this rows the function will select identically the rows of every other block.}
}
\value{
tibble
}
\description{
Takes any NxN data fame and inverse its data block by block based on the predetermination that we'll give to the fonction.
}
