% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F7.R
\name{LargeTrSD}
\alias{LargeTrSD}
\title{SD of transposed large Data block by block}
\usage{
LargeTrSD(
  Data = data(),
  lengthBlock = integer(),
  cls = c(),
  rs = c(),
  SD = character()
)
}
\arguments{
\item{Data}{Data frame that will be imported.}

\item{lengthBlock}{Nomber of columns that will be transposed for any block.}

\item{cls}{The columns of the first block. Based on this columns the function will select identically the rows of every other block.}

\item{rs}{Rows that will define the block that will be transposed.}

\item{SD}{Character that can take the values Row, Col or Block. Based on the option
that we'll choose and the function will return the SD of the Row, Col of each block or the general SD of all the block.}
}
\value{
tibble
}
\description{
Transpose large data frame by block and gives the the SD of each block In case we have a large
data frame and need to transpose it by block this fucntion will transpose it by block than it will give the SD of the column or of the rows of each block or of the
hole block transposed
}
