% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3.R
\name{ProdDtBbB}
\alias{ProdDtBbB}
\title{Product of datas block by block}
\usage{
ProdDtBbB(
  Data1 = data(),
  length_block1 = integer(),
  cl_Dt1 = c(),
  rows1 = c(),
  Data2 = matrix(),
  length_block2 = integer(),
  cl_Dt2 = c(),
  rows2 = c()
)
}
\arguments{
\item{Data1}{First data table or matrix that will be imported to the fonction and used to execute
the matrix production with an other data table/matrix.}

\item{length_block1}{Nomber of rows that will be used to define the lenght of all blocks for
the first data table or matrix.}

\item{cl_Dt1}{The columns of Data1 that we want to include in the first data product process.}

\item{rows1}{he rows of the first block of the Data2. Based on this rows the function will select
identically the rows of every other block.}

\item{Data2}{The second data table or matrix that we will apply the product function block by block
of the first data frame.}

\item{length_block2}{The length of the blocks that will be multiplied by the blocks of the first
data table.}

\item{cl_Dt2}{The columns of Data2 that we want to include in the data product process.}

\item{rows2}{The rows of the first block of  Data2. Based on this rows the function will select
identically the rows of every other block.}
}
\value{
tibble
}
\description{
This function is used for the purpose of having the product block by block of two
different data tables or matrices.
}
