% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{TrDtBbB}
\alias{TrDtBbB}
\title{Transpose data table block by block}
\usage{
TrDtBbB(
  Data = data(),
  lengthBlock = integer(),
  cls = c(),
  rs = c(),
  time_col = NULL
)
}
\arguments{
\item{Data}{Data frame that will be imported}

\item{lengthBlock}{Nomber of rows that will be transposed for any row.}

\item{cls}{Columns of the  data frame that will be included int the block transposetion.}

\item{rs}{The rows of the first block. Based on this rows the function will select identically the rows of every other block.}

\item{time_col}{The 'time parameter' is optional. We can ignore it if our data has no time values.}
}
\value{
A tibble.
}
\description{
This function take any data table or matrix and transpose it block by block.
}
